nlp-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sentence-transformers>=2.2.2
pymilvus>=2.6.0
python-dotenv>=1.0.0
